Update patch set 2
Patch Set 2:

> Not to be philosophical, but to be precise, what exactly does ‟field” refer to here?  Is each individual ‟insert” a field, where that individual field has a ‟Type” property and a ‟Select” property?  Or should ‟field” be understood to refer to the ‟type” of field?

Choosing the exact meaning of "field" term in relation of the agreed upon arbitrary design is what I would encourage you to avoid. Just because it's trying to set in stone (provide a hard terminology) for something random.

So I would *universally* not assign names to fields (so I would not call something "Sender field", or "Name field", or something like that).

In Word, you are free to insert a field with page number, and then use Edit Field, to turn it into a field with, say, ToC. So it's not a "Page Number field", but rather "field with page number", the same way as "this is not a "Lorem Ipsum" paragraph, but a paragraph with "Lorem Ipsum"".

Patch-set: 2
Reviewer: Gerrit User 1000994 <1000994@6622c302-437a-49ad-a91d-55ba7dcf1615>