Update patch set 2
Patch Set 2:

> > But these points only amplify my queries about the current text.  Isn't an implication of your point that the distinction between "Field" type and "Select" (a field) (in the current help) also misleading?
> 
> No. You just somehow implied that "Select" is for "a field", while in fact it was not for "a field", but for "a piece if data".

But that was not me speaking, I was just repeating (and asking about) what is currently written on the help page!

But no matter, it confirms my suspicion that something is wrong there, and reinforces my belief that I understand your point(s).

Thanks for now.  With this new insight, I think the best is that I rewrite those 2 or 3 sentences at the beginning of the help page, and see if it survives your flamethrower. (-:

Patch-set: 2
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>