Update patch set 4
Patch Set 4: Code-Review+1

Alain - many thanks for your comments.  Have I assumed correctly that the changes that I have proposed are correct (both in defining and copying a function in Basic macro)?

From your suggestion, I have now added a link to the "Calling Python Scripts in Basic" page to the "Related Topics" -- a good addition.  As I remember the bug report, the person was searching (in Google) for Python as well.

But I think that I should not mention Python in the steps -- because the Heading on this page (
https://help.libreoffice.org/latest/ug/text/scalc/guide/userdefined_function.html ) is "Defining A Function Using LibreOffice Basic"

Maybe this patch is ready to be pushed?

Patch-set: 4
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=+1