Update patch set 2
Patch Set 2:

The "Glossary" has lost its "sort" ability.  Maybe it is my fault, but I cannot see that I changed anything that is relevant to sort.  The <sort order="asc"> is present, and everything is in <section>

I also tried/tested (in XHPEditor) to clip out all that I had added -- but it still did not sort.

I had updated to <h2> in all the entries -- but they are "inside" of a <section>, so that should not affect the sort?  Or is that the problem?  That DTD for sort does not recognize <h2>?

The present patch is to just to repair the logic of what is linked to what for "register-true" -- but if there is something that I can do to repair the "sort" , then I can include it in this patch.  Thanks.

Patch-set: 2
Reviewer: Gerrit User 1000855 <1000855@6622c302-437a-49ad-a91d-55ba7dcf1615>
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>
Reviewer: Gerrit User 1001613 <1001613@6622c302-437a-49ad-a91d-55ba7dcf1615>
Reviewer: Gerrit User 1000321 <1000321@6622c302-437a-49ad-a91d-55ba7dcf1615>
Work-in-progress: false