Update patch set 3
Patch Set 3:

> Patch Set 3:

> 
> 1.  As best as I can tell, the icons in Navigator Content View are independent of chosen icon theme.
None of icon of LibreOffice independent. Latest UI part which were hard coded was LibreLogo toolbar. But now it is theme-able also 
> 2.  I  finally found an icon that looks like the one in Navigator: icon-themes/breeze/sw/res/sc20179.png 
> 3.  When "sw/res/sc20179.png" is used as "src" (in the .xhp), then it ends up being presented as sc_inserthead.png   (I can see that this "translation" is in the links.txt)  I am pretty sure that this is a mistake.
What's mistake? sc20179.png icon function is to assist cursor to active header in current document. So, sc_insertheader.png is the right target.
> 4.  It is an understandable mistake, but notice that "sc_insertheader" has its top right corner bent down, 

Top right corner bent down was intentional. It is normal metaphor for a paper.

> while the "Headings" icon in Navigator is "flat" across the top.

Header and Heading are two different things obviously

> Here is what you can help me with.
> 
> 5.  Is "sw/res/sc20179.png" the correct icon for Headings in Navigator?

No, sc20179.png is about Header, not Headings
> 6.  If the answer to 5 is "yes", then I hope that I can enter that into the "src" field for images in .xhp files, and the correct icon will appear on help pages.
> 

The answer is "No"
> Thanks for your help.

Patch-set: 3