Update patch set 2
Patch Set 2:

And if the question raises, do not replace the icon directly in source with its current replacement. If later on the missing icon is produced, the Help will pick the new icon, because it will not be in the replacement table anymore.

Patch-set: 2
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>