Update patch set 2
Patch Set 2: Code-Review-1

Ok -- all "available" information is put back.  

NB. I simply asked a question about whether they could be removed.  Now I know the answer.  (-:  

@Adolfo -- I do not know if it is "clutter".  But recently Olivier and I have been discussing that pages should not have too many notes.  And in this case, I naively assumed that because 3.5 was so long ago (and almost at the start), that maybe they were not needed.

(I think, with this particular page , all the functions need to be there, because it is supposed to show the "Mathematical" category in Function Wizard.)

But now that you bring up the question of "appearance".... Take a look at the page in question:  https://help.libreoffice.org/7.1/en-US/text/scalc/01/04060106.html

It just seemed to me that there were already many notes and warning there -- and that the "available" information is given the same prominent presentation as the notes.  If it is as simple a CSS adjustment, then maybe it is a good idea to give them a slightly less prominent appearance.

Nothing critical -- just an explanation of what motivated my naive question.

Meanwhile -- as I understand -- we are waiting to hear from Eike about other functions, so while this patch does what Mike asked, it may be necessary to make additional changes, so -1 for now.

Patch-set: 2
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1