Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

@Luboš, thanks for your helpful comments.
> The Skia options are visible if Skia is available (which is on Windows or on X11 if the gen backend is active). If more backends get Skia support in the future, the options will show for them too.

Thanks for this confirmation.  We have a switch command, so in the future we can include Mac or unix when available.  We decided to not mention X11 (because not commonly used).  

Do you know if there is a recommended or preferred setting?  For example, when I enabled both Skia and "use hardware acceleration" (and then restart), only Skia is enabled and "use hardware acceleration" was disabled and greyed out.  Maybe we can give some information about whether or why a user would choose "hardware acceleration" or "Skia"?

Patch-set: 4
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1