Update patch set 7
Patch Set 7:

I take back that part about merging the two options. While conceptually that would be the correct thing to do, the 'use hardware acceleration' option is technically only for one part of LO (the canvas module) and is tied to it, which makes it complicated to reuse it for Skia. So I'll keep the UI as it is, I've only created https://gerrit.libreoffice.org/c/core/+/106491 to make sure the greying out works correctly.

Patch-set: 7
Reviewer: Gerrit User 1000092 <1000092@6622c302-437a-49ad-a91d-55ba7dcf1615>