Update patch set 3
Patch Set 3:

(1 comment)

If Mike is happy with the content ... then the only remaining issues are (a) Mike's question about the position of <helpdocument version="1.0">  (which I cannot answer), and (b) any opinions about "enable" or "activate" or "choose" for Table > Number Recognition. 

Otherwise..this patch might be ready to go.

Patch-set: 3