Update patch set 6
Patch Set 6:

Looking now at "20-3-4" being converted for you into 2004-03-20 in Dutch locale, and not being converted to a date at all in Russian locale, I suppose that the fallback ISO conversion tries to be smart, and when the string matches Y-M-D, but *possibly might* be also treated as "D-M-Y" (or maybe "M-D-Y"), then it's ignored. Hence 0-1-2 is converted (0 couldn't be day nor month), but 1-2-3 is not (here using Russian locale), since it might be both 2001-02-03, and 2003-02-01... A questionable choice. But maybe there's a rationale for that.

Patch-set: 6
Reviewer: Gerrit User 1000994 <1000994@6622c302-437a-49ad-a91d-55ba7dcf1615>