Update patch set 8
Patch Set 8:

(1 comment)

This version adds an explanation about how to make Date acceptance patterns, and a correction about the Y-M-D pattern.  But as noted in the comment, there is a question about whether that sentence describes the intended LO implementation and whether the implementation works as intended.  This may be only remaining issue to resolve before submitting this patch.

Patch-set: 8
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>