Update patch set 8
Patch Set 8: Code-Review-1

@Luboš -- thanks for clarifications. The description for ‟Force Skia software rendering” was more or less copy+paste from OpenGL (where the control was to force OpenGL to ignore blacklist"), but this case seems to be the opposite, where the hardware is not on the blacklist -- so it is clear that a change is needed.

> This option disables the use of acceleration if a driver is problematic but is not on the list".

Just to be sure that I have this right. Is the following table of combinations correct?  

  ( where "Skia" is always enabled and where "acceleration" =  both "Use hardware acceleration" is enabled and Skia actually uses the graphics driver) (In the new interface, "Use hardware.." is automatically activated (and the control greyed out) when "Use Skia..." is selected, so that is why I write "both" ).

(NB.  Side comment: If "Use Skia.." is deactivated, and "Force" is also checked at the time, then "Force" becomes greyed out, but remains checked -- even though Skia is not checked.  Probably not intended -- and would be confusing, because it would look like "Force Skia software rendering" is being used, when IIUC, it is not.)

1.  "Force" enabled      +  Non-buggy graphics hardware               = Skia and acceleration
2.  "Force" NOT enabled  +  Non-buggy graphics hardware               = Skia and acceleration
3.  "Force" NOT enabled  +  Buggy graphics hardware (not on denylist) = Skia and acceleration

Implication:  Does not matter if "Force" is enabled or not, Skia uses acceleration in these cases.

4.  "Force" NOT enabled  +  Buggy graphics hardware (on denylist)     = Skia w/o acceleration
5.  "Force" enabled      +  Buggy graphics hardware (on denylist)     = Skia w/o acceleration

Implication: If graphics hardware on "denylist" =>  Does not matter if "Force" is checked or not and "Use hardware.." is ignored (even if checked).  (i.e., no possibility for user to control or intervene in cases with driver on denylist)

6. "Force" enabled       +  Buggy graphics hardware (not on denylist) = Skia and acceleration

Implication:  This is the case where the option is relevant to the user -- because the graphics driver is buggy, but not on deny list. 

If this analysis is correct, then wouldn't a label like:  "Use Skia without acceleration"  communicate better what is going on?   And if that is correct, then maybe the "default" (factory) setting would be for that control to be unchecked? 

With that configuration, then the "help" could simply say (something like):  
  Select this control to disable the use of the hardware graphics driver.   

(no need to mention denylists, buggy drivers, etc.,)

Thanks for your consideration.

Patch-set: 8
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1