Update patch set 8
Patch Set 8:

> Just to be sure that I have this right. Is the following table of combinations correct?  
> 
>   ( where "Skia" is always enabled and where "acceleration" =  both "Use hardware acceleration" is enabled and Skia actually uses the graphics driver) (In the new interface, "Use hardware.." is automatically activated (and the control greyed out) when "Use Skia..." is selected, so that is why I write "both" ).

It is just a coincidence that "use hardware ..." is enabled when greyed out by allowing Skia. It could be also disabled, either way it doesn't matter when Skia is enabled.

> 
> (NB.  Side comment: If "Use Skia.." is deactivated, and "Force" is also checked at the time, then "Force" becomes greyed out, but remains checked -- even though Skia is not checked.  Probably not intended -- and would be confusing, because it would look like "Force Skia software rendering" is being used, when IIUC, it is not.)

This is similar to the case above, if the option is greyed out, it doesn't matter. I don't what the UI guidelines are on this, but that's how I understand it and that's how it works internally. I could change the Skia option to get unchecked if needed. I don't know about "use hardware ...", since as I said, that one is tied to the canvas module and if I knew how to control it when Skia is enabled without causing problems, I could just as well reuse it.

> 
> 1.  "Force" enabled      +  Non-buggy graphics hardware               = Skia and acceleration

No. If software rendering is forced, then it is, well, forced. So no acceleration. Your analysis would be correct for the OpenGL option, but not here. If the "Force" option is enabled, then nothing else matters (with the exception of Skia as a whole being disabled).

> 6. "Force" enabled       +  Buggy graphics hardware (not on denylist) = Skia and acceleration
> 
> Implication:  This is the case where the option is relevant to the user -- because the graphics driver is buggy, but not on deny list.

The implication is correct, but not the result above. If the driver is buggy but not listed, then the option allows the user make LO act as if the driver were on the list, so " = Skia w/o acceleration".

> If this analysis is correct, then wouldn't a label like:  "Use Skia without acceleration"  communicate better what is going on?   And if that is correct, then maybe the "default" (factory) setting would be for that control to be unchecked?

Well, yes. As I said before, the option is really the opposite of the "use hardware acceleration" option, it's just that each of them apply to different piece of code. If the dialog code for that option wasn't so tied to its limited meaning, I'd reuse it. That could possibly be fixed, but it'd require a non-trivial amount of work. So, in other words, we don't live in a perfect world.

> With that configuration, then the "help" could simply say (something like):  
>   Select this control to disable the use of the hardware graphics driver.   

Yes, that's a good description of what the "force sw rendering" option is for.

Patch-set: 8
Reviewer: Gerrit User 1000092 <1000092@6622c302-437a-49ad-a91d-55ba7dcf1615>