Update patch set 9
Patch Set 9:

@Luboš 

> It is just a coincidence that "use hardware ..." is enabled when greyed out by allowing Skia. It could be also disabled, either way it doesn't matter when Skia is enabled.

I leave the interface decision to you.  You mentioned that "Use hardware" has some consequences when for "canvas" that might be separate from Skia.  (FWIW, the present "coincidence" serves to give a correct impression that the graphics hardware is being used, even if Skia's use is internally separate from "Use hardware..." - but I think that we agree that we do not need to mention/explain this in the help.)

> > (NB.  Side comment: If "Use Skia.." is deactivated, and "Force" is also checked at the time, then "Force" becomes greyed out, but remains checked -- even though Skia is not checked.  Probably not intended -- and would be confusing, because it would look like "Force Skia software rendering" is being used, when IIUC, it is not.)
> 
> This is similar to the case above, if the option is greyed out, it doesn't matter. I don't what the UI guidelines are on this, but that's how I understand it and that's how it works internally. I could change the Skia option to get unchecked if needed. 

I think it would be a good thing. Otherwise from user POV, the check makes it look like "Force Skia.." is still activated (even though the control is disabled).  [No consequences for the help page.]

> Well, yes. As I said before, the option is really the opposite of the "use hardware acceleration" option, it's just that each of them apply to different piece of code. If the dialog code for that option wasn't so tied to its limited meaning, I'd reuse it. That could possibly be fixed, but it'd require a non-trivial amount of work. So, in other words, we don't live in a perfect world.

(-:   Maybe, at least, we have succeeded in giving a reasonable documentation for it on the help page.

If you are happy with the current formulations, then this patch should be ready to go.

Patch-set: 9
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>