Update patch set 2
Patch Set 2:

> The change looks like missing changed files...

@Mike - now it should be all right. 

NB. I discovered your tdf#130738 after I made this patch - so the focus was only on adding the Custom condition.  But I can see already that this should probably be improved -- and I trust you will have some suggestions. (-:

Also, I could not fully understand tdf#131309 (it was unclear to me if it was actually a bug, or just a misunderstanding of how Custom works -- and that maybe some documentation could help in relation to that point?)

Patch-set: 2
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>