Update patch set 4
Patch Set 4:

(1 comment)

Sorry -- one more -- Look in "Opening Documents With Templates" section in: 

https://help.libreoffice.org/7.2/en-US/text/shared/01/01020000.html

The comment marks the relevant xml code. 

A couple of questions.  

1.  It seems like subdirectories are also used  (not documented on help page). So -- choosing C:/Users/username/ will look in all subdirectories  (this belief comes from seeing how long it takes the Template Manager to open Ctrl+Shift+N, also all my subdirectories under "username" become Categories in the Template Manager!).  - so maybe need to add a "note"  about care in choosing a top-level directory.

2.  Internal paths are also shown in Options, so I am thinking a simple way to refer to "shared templates" is to say "look in Internal Paths"  

(aren't these actually what is also called "built-in templates" ? seems better to call them that, to be consistent with other help pages on templates)

Patch-set: 4