Update patch set 5
Patch Set 5:

> The {user} there should be universally replaced. It's misleading. It might be changed to e.g. {profile}.

I think that {user} is being used a "placeholder" -- not as a "literal".  Look at the sentence just before the table, where you will see:

"user data for each user is stored in the {user} directory, which is located in the user's Documents and Settings directory."

and then the sys-switch is used to tell what it should be for UNIX and default. 

Do these facts change your opinion?

Patch-set: 5