Update patch set 5
Patch Set 5:

> I think that {user} is being used a "placeholder" -- not as a "literal".  Look at the sentence just before the table, where you will see:

I know what it is used for. And it is wrong. Because it is misleading, having a literal "user" subdirectory; and further, there is a term used for this thing, which should be used consistently here, and that term is "user profile".

Patch-set: 5
Reviewer: Gerrit User 1000994 <1000994@6622c302-437a-49ad-a91d-55ba7dcf1615>