Update patch set 2
Patch Set 2:

Adolfo -- a practical question -- this patch is to update the help page for the Template Manager.
I was waiting to start its review for different reasons -- but now I guess we are close to the string freeze -- plus -- in 7.2 -- a listview was added to the template manager, which I was going to document here -- but then this would make the patch useless in relation to 7.1.

The present state of this patch is mostly to correct some errors, plus I have added icons.  (all details in commit message).  It is "complete" in the sense that everything is as intended.  I was just planning to develop it further.

My question:  Should I submit this patch now -- so that it can be used in relation to 7.1?  Or should I just continue and develop it in relation to 7.2.  I have no opinion. Just trying to accommodate.

Patch-set: 2
Reviewer: Gerrit User 1000855 <1000855@6622c302-437a-49ad-a91d-55ba7dcf1615>
Reviewer: Gerrit User 1000321 <1000321@6622c302-437a-49ad-a91d-55ba7dcf1615>
Work-in-progress: false