Update patch set 4
Patch Set 4: Code-Review-1

When adding a new help file, the file must be listed in the proper gbuild makefile, to be picked for build.

The makefile to change and to add to this patch is AllLangHelp_scalc.mk in the root of the Help submodule. Have a look at this file and add the new file name at the proper place, respecting alphabetical order (so to speak).

Patch-set: 4
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1