tdf#139122 Add information about locking/unlocking toolbars
The toolbars in the default "Standar Toolbar" UI are locked by default
since 7.0, so "Using Toolbars" help page is updated to reflect this
change and information about locking and unlocking toolbars are added.

Also update the entire page to use new XML tags, and some general
clean-up.

+ add a paragraph at the beginning of the page explaining the
  docked/floating status of toolbars, and mention that toolbars are
  docked and locked by default
+ add two sections about unlocking and locking docked toolbars and
  corresponding bookmark
* move description of toolbar handle from undocking section to unlocking
  section
- remove unnecessary list structure in the section about undocking
  toolbars
* use more specific XML tags throughout the page, such as <h1><h2><h3>
  tags instead of <paragraph> for headings, as well as <warning><note>
  tags instead of <paragraph>, and <menuitem> tags instead of <emph>,
  where appropriate.
* use en-dash instead of ASCII hyphen for menu items
- delete old comments intended as changelogs and no longer relevant

The "Part of the main window - Toolbars" section of _Getting Started
Guide 6.4_ (Chapter 1, pages 20-22) is used as reference.

Change-Id: Id6256c75cd99e7b7db69bbff2a4d36be72a2d7cf