tdf#139122 Add information about locking/unlocking toolbars
The toolbars in the default "Standar Toolbar" UI are locked by default
since 7.0, so "Using Toolbars" help page is updated to reflect this
change and information about locking and unlocking toolbars are added.

+ add a paragraph at the beginning of the page explaining the
  docked/floating status of toolbars, and mention that docked toolbars
  can be locked
+ add two sections about unlocking and locking docked toolbars and
  corresponding bookmark
* move description of toolbar handle from undocking section to unlocking
  section
* use <h2> tags for headings in new and modified sections
- remove unnecessary list structure in the section about undocking
  toolbars
- delete an old comment intended as a changelog and no longer relevant

The "Part of the main window - Toolbars" section of _Getting Started
Guide 6.4_ (Chapter 1, pages 20-22) was used as reference.

Change-Id: Id6256c75cd99e7b7db69bbff2a4d36be72a2d7cf