Update patch set 2
Patch Set 2:

I am still new to XHP help file format, so I mostly adopted the old structures and tags used in the original file, and didn't use the new and better tags, such as for headings and menuitems.

Also I'm not a native speaker, so corrections about grammar and wordings are very welcome.

Patch-set: 2
Work-in-progress: false