Update patch set 13
Patch Set 13:

Thanks Ilmari and Mike -- but your comments about "share" directory have referring to ps 7, which was no longer current.  I had found a solution where it was not necessary to refer to "share" directory (because this help page is about paths that users can edit).  

But now, in response to your comments, I have added an explicit part about the "share" directory (which was not explained in the original page) at the end of this page  (and a few more updates on help pages about editing paths).  The current version is now ps 13

Patch-set: 13
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>