Update patch set 5
Patch Set 5:

(3 comments)

Now that the "syntactical" questions are out of the way --- can pay more attention to the "structural" ones... as you already noted -- the organization of the page could be improved. Have offered some thoughts for consideration -- plus a possible way to handle the toolbar handle variations.

Patch-set: 5