Update patch set 6
Patch Set 6:

(3 comments)

Made another round of changes according to Seth's suggestions.

I didn't address the issue about combining the two sections of unlocking (and positioning) toolbars and of making toolbars floating.  I feel the structure of this page has always been about pairing the actions -- open/close, dock/undock, and now lock/unlock.  Combining unlocking and undocking into one section would break this structure, and to me it's not worth it.  I think the current wording is clear enough for a user to navigate and find what he/she needs.

I don't think I'll put more efforts on changing the structure of the page.  Those efforts would be better spent on rewriting the whole page.  My goal is just to add information about the current "locked by default" status and how to unlock/lock toolbars.  I only modified the undocking section because it needs unlocking as a prerequisite and it refers to the toolbar handle which are invisible by default now.

Of course, a more competent writer is welcome to further improve this patch.  I guess we are in string freeze for 7.1 anyway and there is no rush.

Patch-set: 6