Update patch set 7
Patch Set 7:

> I also changed <emph> to <menuitem> in the whole page, and used en-dash instead of ASCII hyphen, to make everything consistent.  Of course this would trigger re-translation, so these can be reverted if necessary.  Though I don't think it's a big deal at this point of a development cycle, unless we plan to backport this to 7-1 branch as well.  

Probably all right.  Some translators get irritated at having to "redo" entire paragraphs, because of a change from emph to menuitem (for example).   (Lately, I have only made those kinds of changes when I am also changing the text in the paragraph -- because emph and menuitem appears the same).

> (What exactly is the string freeze policy for help documentation?)
The exact same as with UI changes.  At this point, no help changes are sent to 7-1 branch, except for broken links (because they are not strings that need translation)

> Two more small things remain:
> 1. There is still one place that talk about "Font Color" icon that uses <emph>, should it be changed to <widget>?  Or is <emph> appropriate here?
widget puts a "box" around the word -- making it look like a "button", so I use them with buttons in dialog boxes.  For icons on a toolbar (as in this case), then I use emph.

> 2. There are three more old comments on the page, all look like used as changelog, and are probably not relevant anymore.  Delete them all?
Yeah, they do not seem so relevant anymore.  The numbers that are given do not refer to tdf bug numbers.

Patch-set: 7