Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (2 comments)
> 
> Hi Martin - are these two inlineswitches (see comments) going to create translating problems for you?

It is string freeze for LO71, I guess you and the rest of documentation team understand what that means.

Help does not change during string freezes and all help updates go to master ...

If there is something burning necessary to change/update, it must be communicated with Sophie, who is the l10n lead - and if that change is accepted - it must be communicated to all l10n teams so they can localize the updated string(s) in time for the release and so they are aware that there have been string changes after the string freeze.

This it the only way that l10n teams can make fully localizable versions in time for the release.

So I am not the one to be asked, I can translate these strings in 10 minutes. But I cannot judge for all 100+ l10n teams.

Please write to the l10n list and explain why this patch is necessary to be included in LO71.

Patch-set: 5