Update patch set 9
Patch Set 9:

> Right, so the handle is not necessarily vertical.  And here I thought "at the start of the toolbar" was just taking RTL user interface into consideration...

@Ming - Not likely to be an issue...here is why:

To start the handle is "vertical" and the page is aimed at users who are not familiar with docking/unlocking. So better that the page gives some meaning for someone who is trying to figure this out for the first time.  If someone has been able to make "horizontal toolbars", then it is likely that the help page is being used anymore, and they understand the purpose of the icon at the start of the toolbar. 

As you already noted, there are some aspects that could be better in that page, and now that you are "familiar" with the page -- maybe you will come back to it someday--

For example -- maybe the example with Font Color is wrong.  A dialog box opened, but it is not a toolbar in the sense that it can be docked (or shown with View - Toolbars), or locked.  (or maybe it IS supposed to be a toolbar and it is a bug -)  From the first point of view, the sentence should probably be deleted, from the "bug" point of view, then it should be left alone.

Also, afaict, only Font Color and Highlight Color work like this.  Maybe this point should appear at the end of this page, because it is not really about docking and locking toolbars, but a point about using some icons on a toolbar.

And the context menus are a little more tricky -- because it presupposes that it has not been unchecked in View - Toolbars.

Patch-set: 9