Update patch set 3
Patch Set 3:

Comment 6 in tdf#139286 made me aware of tdf#137347 with a similiar updating need, which is now also handled in this patch.  And added "Slides Pane" to the header, because same shortcuts apply, except for "Enter", which is noted.
Screenshot of this version is at: http://bugs.documentfoundation.org/attachment.cgi?id=168556

Patch-set: 3