Update patch set 2
Patch Set 2: Code-Review-1

Some remarks.

There is an excessive use of bold (<emph>) characters in the text, some terms can be set as normal char to reduce reading fatigue. For example you can remove the bold character of "Template manager" when writing about it in a sentence. Same for 'Title', 'internal path', 'user profile', Categories, user path...

An entry in the shared.tree ('Contents' on the right) is welcome, in "Common Help Topics/Guides". Otherwise there is no clear way to navigate to this page.

Bookmarks can be improved: "template manager;how to use"

The sentence "If two templates have the same entry in the Title field and belong to the same Category, then only one template appears in the Template Manager." should be in a <note>.

"groupuinames.xml": This kind of info is not recomended for end-users. They should not touch such files or a call to help desk is to be expected $$$.

The warning at the end is very hard to understand. Can you reword it?

D'oh! You found a bug (text/shared/guide/template_manager.xhp#manage_templates not found).

Patch-set: 2
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1