Update patch set 3
Patch Set 3:

> Some remarks.

Thanks for feedback.  Others are also welcome to comment.

> There is an excessive use of bold (<emph>) characters in the text, some terms can be set as normal char to reduce reading fatigue. For example you can remove the bold character of "Template manager" when writing about it in a sentence. Same for 'Title', 'internal path', 'user profile', Categories, user path...

Done 

> An entry in the shared.tree ('Contents' on the right) is welcome, in "Common Help Topics/Guides". Otherwise there is no clear way to navigate to this page.

Was planning to add a link in "Related Topics" for "Template Manager" (after this was submitted).

In principle, a user should only use Template Manager.  This page is meant for users who want to go beyond the UI to understand what is behind, so that they can manipulate the filesystem directly, rather than go through the Template Manager.  In this sense, it is not really a "guide" about "normal" or "recommended" use.

> Bookmarks can be improved: "template manager;how to use"

This is not really a guide to using the template manager.  There is another page about the template manager (and a patch to update it, https://gerrit.libreoffice.org/c/help/+/107549 , waiting to be submitted).

There is no intention to encourage users to know/use this information, but these issues come up from time to time in BZ -- which is what motivated tdf#122720. 

> The sentence "If two templates have the same entry in the Title field and belong to the same Category, then only one template appears in the Template Manager." should be in a <note>.

Done.

> "groupuinames.xml": This kind of info is not recomended for end-users. They should not touch such files or a call to help desk is to be expected $$$.

As Mike K. has argued in relation to the "paths" page, some information in help can also be relevant for system administrators.  

Also, as noted before, this page is meant for people who want to go beyond the template manager.  It is not meant as a guide to use the template manager, but it is necessary for understanding the relation between categories and subdirectories.

> The warning at the end is very hard to understand. Can you reword it?

Moved one piece to a separate <note> and modified what remained.

> D'oh! You found a bug (text/shared/guide/template_manager.xhp#manage_templates not found).

The plan is to add this present patch after the patch that adds the relevant variable for template manager ( https://gerrit.libreoffice.org/c/help/+/107549 ) is submitted.

Patch-set: 3