Update patch set 6
Patch Set 6: Code-Review+1

Sorry -- had to correct a merge conflict -- everything should be ready to go. Have corrected mistakes in the "Attributes" page, and corrected a mistake in the "Find and Replace" , moved two notes that were next to each other, and improved the hierarchy for the controls so that it is easier to read the page.  As mentioned from the beginning -- it looks good to me -- but probably worth another having a look.

Patch-set: 6
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=+1