Update patch set 2
Patch Set 2:

Hi Ilmari -- no rush - but I would be grateful, when you get a chance, to take a look at a few parts of the content changes that I made to this help page.

Commit message explains the background for the changes. 

The interesting parts for review are:
(a) at the beginning (just after "how to get") where I tried to clarify points that are raised in the bug reports.   (Does the content seem ok?)

(b) the "Apply Style" section
In my experiments with Apply Style, I could sometimes get a heading without starting with an uppercase letter, but it worked reliably with uppercase, which is why I put it that way in the help.
(cool feature)  (Anyway, does the text seem ok?  e.g., does it make it easy for you to use this feature? )

Patch-set: 2
Reviewer: Gerrit User 1000855 <1000855@6622c302-437a-49ad-a91d-55ba7dcf1615>
Reviewer: Gerrit User 1001613 <1001613@6622c302-437a-49ad-a91d-55ba7dcf1615>
Work-in-progress: false